Refactor config header and uptime symbol.
Move 'config.h' to 'pico-cec/config.h' as an initial attempt to
namespace the header.
Add the 'util' module to house very generic symbols (eg. uptime).
