+++ include/pico-cec/util.h
@@ -1,0 +1,8 @@
+#ifndef PICO_CEC_UTIL_H
+#define PICO_CEC_UTIL_H
+
+#include <stdint.h>
+
+uint64_t util_uptime_ms(void);
+
+#endif
